backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
pymysql==1.1.0
python-dotenv==1.0.0
pydantic==2.9.0
python-multipart==0.0.6
faker==20.1.0
gtts==2.4.0
